+++ tests/typing/check_flatten.py
@@ -10,15 +10,13 @@ from orcaset import (
     Effect,
     Period,
     Series,
-    accrue,
     continue_series,
-    covered,
     date_split,
-    exact,
     get,
     period_split,
 )
 from orcaset.maybe import Maybe, Na, isna
+from orcaset.query import accrue, covered, exact
 
 Q1 = Period(date(2025, 1, 1), date(2025, 4, 1))
 Q2 = Period(date(2025, 4, 1), date(2025, 7, 1))